funções: habitos de hoje marcar e desmarcar um hábito, e barra de porcentagem

+++ src/components/Habit/index.js
@@ -37,19 +37,23 @@ const Habit = ({ habit, token, setLoading, getHabitList }) => {
   }, [])
 
   function handleDeleteClick () {
-    setLoading(true)
-    const promise = deleteHabit(habit.id, token);
-    promise.catch((error) => {
-      console.log(error);
-      alert(`Ocorrey um erro: ${error.message}`)
-      setLoading(false);
-      getHabitList();
-    })
-    promise.then(res => {
-      console.log(res.data);
-      setLoading(false);
-      getHabitList();
-    })
+    const confirm = window.confirm('Deseja deletar este hábito?');
+
+    if(confirm){
+      setLoading(true)
+      const promise = deleteHabit(habit.id, token);
+      promise.catch((error) => {
+        console.log(error);
+        alert(`Ocorrey um erro: ${error.message}`)
+        setLoading(false);
+        getHabitList();
+      })
+      promise.then(res => {
+        console.log(res.data);
+        setLoading(false);
+        getHabitList();
+      })
+    }
   }
 
   return (
